renames index.tsx to component name
