+++ src/signals/incident/components/IncidentPreview/components/ContainerListPreview/ContainerListPreview.tsx
@@ -1,14 +1,12 @@
 import React from 'react';
 import type { FeatureType, Item } from 'signals/incident/components/form/ContainerSelect/types';
-import ContainerList from '../../../form/ContainerSelect/components/ContainerList';
+import ContainerList from '../../../form/ContainerSelect/components/ContainerList/ContainerList';
 
 export interface ContainerListPreviewProps {
   value: Item[];
   featureTypes: FeatureType[];
 }
 
-const ContainerListPreview: React.FC<ContainerListPreviewProps> = ({ value, featureTypes }) => (
-  <ContainerList selection={value} featureTypes={featureTypes} />
-);
-
+const ContainerListPreview: React.FC<ContainerListPreviewProps> = ({ value, featureTypes }) =>
+  <ContainerList selection={value} featureTypes={featureTypes} />;
 export default ContainerListPreview;
