+++ src/signals/incident/components/form/ContainerSelect/ContainerSelect.tsx
@@ -1,0 +1,64 @@
+import type { FunctionComponent } from 'react';
+import React, { useCallback, useState } from 'react';
+import { ContainerSelectProvider } from './context';
+import Intro from './components/Intro/Intro';
+import Selector from './components/Selector';
+import Summary from './components/Summary';
+import type { ClickEventHandler, Item, Meta } from './types';
+import type { Incident } from 'types/incident';
+import type { LatLngExpression } from 'leaflet';
+
+export interface ContainerSelectProps {
+  handler: () => { value: Item[] };
+  meta: Meta;
+  parent: {
+    meta: {
+      incidentContainer: { incident: Pick<Incident, 'location'> };
+      updateIncident: (data: { extra_container: Item[] }) => void;
+    };
+  };
+}
+
+const ContainerSelect: FunctionComponent<ContainerSelectProps> = ({ handler, meta, parent }) => {
+  const value = handler().value;
+  const [showMap, setShowMap] = useState(false);
+
+  const { coordinates } = parent.meta.incidentContainer.incident.location.geometrie;
+  const location: LatLngExpression = [coordinates[1], coordinates[0]];
+
+  /* istanbul ignore next */
+  const update = useCallback(
+    (selectedValue: Item[]) => {
+      parent.meta.updateIncident({ extra_container: selectedValue });
+    },
+    [parent]
+  );
+
+  const edit = useCallback<ClickEventHandler>(
+    event => {
+      event.preventDefault();
+      setShowMap(true);
+    },
+    [setShowMap]
+  );
+
+  const close = useCallback<ClickEventHandler>(
+    event => {
+      event.preventDefault();
+      setShowMap(false);
+    },
+    [setShowMap]
+  );
+
+  return (
+    <ContainerSelectProvider value={{ selection: value, location, meta, update, edit, close }}>
+      {!showMap && value.length === 0 && <Intro />}
+
+      {showMap && <Selector />}
+
+      {!showMap && value.length > 0 && <Summary />}
+    </ContainerSelectProvider>
+  );
+};
+
+export default ContainerSelect;
