+++ src/signals/incident/components/form/ContainerSelect/components/Intro/index.tsx
@@ -1,0 +1,3 @@
+import Intro from './Intro';
+
+export default Intro;
