+++ src/signals/incident/components/form/ContainerSelect/components/Selector/components/LegendToggleButton/__tests__/LegendToggleButton.test.tsx
@@ -6,7 +6,7 @@ import { render, screen } from '@testing-library/react';
 import { withAppContext } from 'test/utils';
 
 import LegendToggleButton from '..';
-import type { LegendToggleButtonProps } from '..';
+import type { LegendToggleButtonProps } from '../LegendToggleButton';
 import { MapPanelProvider } from '@amsterdam/arm-core';
 import { SnapPoint } from '@amsterdam/arm-core/lib/components/MapPanel/constants';
