+++ src/signals/incident/components/form/ContainerSelect/components/Selector/components/WfsLayer/__tests__/WfsLayer.test.tsx
@@ -11,7 +11,7 @@ import containersJson from 'utils/__tests__/fixtures/containers.json';
 import MAP_OPTIONS from 'shared/services/configuration/map-options';
 import type { DataLayerProps } from '../../../../../types';
 import WfsDataContext, { NO_DATA } from '../context';
-import WfsLayer, { isLayerVisible } from '..';
+import WfsLayer, { isLayerVisible } from '../WfsLayer';
 
 const fetchMock = fetch as FetchMock;
