+++ src/signals/incident/components/form/ContainerSelect/components/Selector/components/WfsLayer/components/ContainerLayer/__tests__/index.test.tsx
@@ -21,12 +21,10 @@ const options: MapOptions = {
   zoom: 14,
 };
 
-const withMapContainer = (Component: ReactNode) => (
+const withMapContainer = (Component: ReactNode) =>
   <Map data-testid="map-test" options={options}>
     {Component}
-  </Map>
-);
-
+  </Map>;
 describe('ContainerLayer', () => {
   const featureTypes: FeatureType[] = [
     {
@@ -54,11 +52,10 @@ describe('ContainerLayer', () => {
   ];
 
   it('should render the cluster layer in the map', () => {
-    const ContainerLayerWrapper = () => (
+    const ContainerLayerWrapper = () =>
       <WfsDataProvider value={containersJson as FeatureCollection}>
         <ContainerLayer featureTypes={featureTypes} />;
-      </WfsDataProvider>
-    );
+      </WfsDataProvider>;
     render(withMapContainer(<ContainerLayerWrapper />));
 
     expect(screen.getByTestId('map-test')).toBeInTheDocument();
