+++ src/signals/incident/components/form/ContainerSelect/components/Summary/__tests__/Summary.test.js
