+++ src/signals/incident/components/form/ContainerSelect/components/Summary/index.tsx
@@ -1,0 +1,3 @@
+import Summary from './Summary';
+
+export default Summary;
